Filter deprecated kubenetes templates
With setting: `generate_tests_kubernetes_test_cases_latest_only: true`
if the latest kubernetes template is deprecated then no kubernetes test
case is executed. This behaviour is confusing because testing a
deprecated template is surely never the intention (they are not
displayed to the Azimuth user)?

Instead, the proposed change ensures that only the latest,
non-deprecated template is selected for testing.